# backend/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: yourfriends@flexiwan.com
    name: API support
    url: https://flexiwan.com
  description: This is the REST API for flexiWAN management.
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://flexiwan.com/legal/terms-of-service/
  title: FlexiWAN API
  version: 1.0.0
servers:
- description: Main server
  url: https://manage.flexiwan.com:443/api
paths:
...
  /devices:
    get:
      operationId: devicesGET
      parameters:
      - description: The number of items to skip before starting to collect the result
          set
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
        example: 10
      - description: Type of returned data
        explode: true
        in: query
        name: response
        required: false
        schema:
          enum:
          - detailed
          - summary
          - ids
          type: string
        example: summary
      - description: The number of items to return, total records count added to the 'records-total' header of the response
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
        example: 10
      - description: The field by which the data will be ordered
        explode: true
        in: query
        name: sortField
        required: false
        schema:
          type: string
        style: form
        example: name
      - description: Sorting order
        explode: true
        in: query
        name: sortOrder
        required: false
        schema:
          type: string
        style: form
        example: asc
      - description: Stringified array of filter objects
        explode: true
        in: query
        name: filters
        required: false
        schema:
          type: string
        style: form
        example: '[{"key":"name","op":"contains","val":"hub"}]'
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Device'
                type: array
                x-content-type: application/json
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Get all registered devices
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
    delete:
      operationId: devicesDELETE
      parameters:
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicesDeleteRequest'
      responses:
        "204":
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - bearerAuth: []
      - apiKey: []
      summary: Delete devices by IDs or matching the filter
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
...
  /jobs:
    get:
      operationId: jobsGET
      parameters:
      - description: A filter on the job status
        explode: true
        in: query
        name: status
        required: true
        schema:
          enum:
          - all
          - complete
          - failed
          - inactive
          - delayed
          - active
          type: string
        style: form
      - description: The number of items to skip before starting to collect the result
          set
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
        example: 10
      - description: The numbers of items to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
        example: 10
      - description: Stringified array of filter objects
        explode: true
        in: query
        name: filters
        required: false
        schema:
          type: string
        style: form
        example: '[{"key":"data.message.title","op":"contains","val":"upgrade"}]'
      - description: A filter on specific job ids (comma separated)
        explode: true
        in: query
        name: ids
        required: false
        schema:
          type: string
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Job'
                type: array
                x-content-type: application/json
          description: Successful operation
        "400":
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - bearerAuth: []
      - apiKey: []
      summary: Get all jobs
      tags:
      - Jobs
      x-swagger-router-controller: Jobs
      x-openapi-router-controller: JobsController
      x-openapi-router-service: JobsService
    delete:
      operationId: jobsDELETE
      parameters:
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsDeleteRequest'
      responses:
        "204":
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Delete jobs by IDs or matching the filter
      tags:
      - Jobs
      x-swagger-router-controller: Jobs
      x-openapi-router-controller: JobsController
      x-openapi-router-service: JobsService
...
components:
...
  schemas:
...
    JobsDeleteRequest:
      properties:
        ids:
          items:
            type: integer
          type: array
        filters:
          items:
            $ref: '#/components/schemas/Filter'
          type: array
...
    Job:
      example:
        progress: 1
        created_at: created_at
        _id: 0
        state: 5
        type: type
        priority: 6
        error: error
        attempts:
          max: 7
          made: 5
          remaining: 2
      properties:
        _id:
          type: integer
        type:
          type: string
        data:
          type: object
        result:
          type: object
        priority:
          type: integer
        error:
          type: string
        progress:
          type: string
        state:
          type: string
        created_at:
          type: string
        attempts:
          type: object
      type: object
...
    Device:
      example:
        interfaces:
        - IPv6: IPv6
          PublicIP: PublicIP
          PublicPort: PublicPort
          NatType: NatType
          useStun: true
          useFixedPublicPort: false
          monitorInternet: true
          internetAccess: 'yes'
          IPv4: IPv4
          type: WAN
          gateway: gateway
          metric: 0
          mtu: 1500
          MAC: MAC
          dhcp: "no"
          routing: OSPF
          IPv6Mask: "64"
          isAssigned: true
          driver: driver
          IPv4Mask: "24"
          name: name
          devId: devId
          _id: _id
          pathlabels: path labels list
        - IPv6: IPv6
          PublicIP: PublicIP
          PublicPort: PublicPort
          NatType: NatType
          useStun: true
          useFixedPublicPort: false
          monitorInternet: true
          internetAccess: 'no'
          IPv4: IPv4
          type: WAN
          gateway: gateway
          metric: 0
          mtu: 1500
          MAC: MAC
          dhcp: "no"
          routing: OSPF
          IPv6Mask: "64"
          isAssigned: true
          driver: driver
          IPv4Mask: "24"
          name: name
          devId: devId
          _id: _id
          pathlabels: path labels list
        org: org
        isConnected: true
        upgradeSchedule:
          jobQueued: true
          _id: _id
          time: "2019-12-19T12:10:01.667Z"
        description: description
        deviceStatus: deviceStatus
        deviceToken: deviceToken
        labels:
        - labels
        - labels
        machineId: machineId
        site: site
        hostname: host01
        versions:
          agent: agent
          router: router
          device: device
          vpp: vpp
          frr: frr
        name: name
        _id: _id
        isApproved: true
        applications:
        - app: app
          status: status
        staticroutes:
        - id: 1
          destination: 20.20.0.0
          gateway: 20.20.0.1
          interface: eth0
        - id: 2
          destination: 10.10.0.0
          gateway: 10.10.0.1
          interface: eth1
        deviceSpecificRulesEnabled: true
        firewallRules:
        - description: description
          direction: outbound
          enabled: true
          action: allow
        - description: description
          direction: inbound
          enabled: false
          action: deny
        fromToken: fromToken
        account: account
        ipList: ipList
        dhcp: dhcp list
        coords:
        - 40.4
        - 5.7
        sync: device sync status
      properties:
        _id:
          maxLength: 24
          minLength: 24
          type: string
        machineId:
          type: string
        name:
          type: string
        description:
          type: string
        site:
          type: string
        isApproved:
          type: boolean
        isConnected:
          type: boolean
        hostname:
          type: string
        deviceStatus:
          type: object
        deviceToken:
          type: string
        fromToken:
          type: string
        account:
          type: string
        org:
          type: string
        ipList:
          type: string
        staticroutes:
          items:
            $ref: '#/components/schemas/StaticRoute'
          type: array
        dhcp:
          items:
            $ref: '#/components/schemas/Dhcp'
          type: array
        coords:
          items:
            type: number
          type: array
        upgradeSchedule:
          $ref: '#/components/schemas/UpgradeSchedule'
        labels:
          items:
            type: string
          type: array
        interfaces:
          items:
            $ref: '#/components/schemas/Interface'
          type: array
        deviceSpecificRulesEnabled:
          type: boolean
        firewallRules:
          items:
            $ref: '#/components/schemas/FirewallRule'
          type: array
        versions:
          $ref: '#/components/schemas/DeviceVersion'
        sync:
          $ref: '#/components/schemas/DeviceSyncStatus'
        pendingNotifications:
          type: integer
      type: object
    UpgradeSchedule:
      example:
        jobQueued: true
        _id: _id
        time: "2019-12-19T12:10:01.667Z"
      properties:
        _id:
          maxLength: 24
          minLength: 24
          type: string
        time:
          type: string
          nullable: true
        jobQueued:
          type: boolean
      type: object
...
    DeviceVersion:
      example:
        agent: agent
        router: router
        device: device
      properties:
        agent:
          type: string
        device:
          type: string
        router:
          type: string
      type: object
...
    DeviceSyncStatus:
      example:
        state: sync state
        trials: remaining automatic sync trials
        autoSync: auto sync state
      properties:
        state:
          enum:
          - synced
          - syncing
          - not-synced
          - unknown
          type: string
        autoSync:
          enum:
          - on
          - off
          type: string
        trials:
          type: number
          minimum: 0
          maximum: 3
      required:
        - state
        - autoSync
        - trials
...
    StaticRoute:
      example:
        destination: destination
        ifname: interface
        gateway: gateway
        metric: metric
        status: "completed"
        redistributeViaOSPF: true
      properties:
        _id:
          type: string
        destination:
          description: Destination
          maxLength: 24
          type: string
        gateway:
          description: Gateway to use
          maxLength: 24
          type: string
        ifname:
          description: Interface name
          maxLength: 24
          type: string
        metric:
          description: Route metric
          type: string
        status:
          description: Route status
          type: string
        redistributeViaOSPF:
          description: Indicates whether the path is redistributed using OSPF.
            Note that the "true" value will return if in the OSPF Global Settings, the redistribution option is enabled for all static routes
          type: boolean
      type: object
...
    Dhcp:
      example:
        _id: "5e5f662f1b0eba50742b6223"
        interface: "0000:00:08.00"
        rangeStart: "20.20.20.2"
        rangeEnd: "20.20.20.255"
        dns: ["8.8.8.8", "8.8.8.4"]
        macAssign: [{"host":"host1", "mac":"08:00:27:fd:00:00", "ipv4":"20.20.20.20"}]
        status: "completed"
      properties:
        _id:
          maxLength: 24
          minLength: 24
          type: string
        interface:
          description: Interface bus Address
          maxLength: 24
          type: string
        rangeStart:
          description: Start IP Range
          maxLength: 20
          type: string
        rangeEnd:
          description: End IP Range
          maxLength: 20
          type: string
        dns:
          description: List of DNS address
          items:
            type: string
            maxLength: 20
          type: array
        macAssign:
          description: List of MAC Assignments
          items:
            $ref: '#/components/schemas/DhcpMacAssign'
          type: array
        status:
          description: Route status
          type: string
      type: object
...
    FirewallRule:
      example:
        _id: id
        description: description
        priority: priority
        enabled: true
        direction: inbound
        inbound: edgeAccess
        classification: classification
        action: allow
        interfaces: interfaces
      properties:
        _id:
          maxLength: 24
          minLength: 24
          type: string
        description:
          maxLength: 100
          type: string
        priority:
          type: integer
          minimum: 0
          maximum: 1000
        enabled:
          type: boolean
        direction:
          enum:
          - inbound
          - outbound
          type: string
        inbound:
          enum:
          - edgeAccess
          - portForward
          - nat1to1
          type: string
        classification:
          schema:
          $ref: '#/components/schemas/FirewallRuleClassification'
        action:
          enum:
          - allow
          - deny
          type: string
        internalIP:
          type: string
        internalPortStart:
          type: string
        interfaces:
          items:
            type: string
          type: array
      required:
      - _id
      - priority
      - enabled
      - direction
      - classification
      - action
      type: object
...
    Error:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    DevicesDeleteRequest:
      properties:
        ids:
          items:
            type: string
          type: array
        filters:
          items:
            $ref: '#/components/schemas/Filter'
          type: array
    Filter:
      example:
        key: name
        op: ==
        val: example
      properties:
        key:
          type: string
        op:
          type: string
        val:
          type: string
      required:
      - key
      - op
      - val
      type: object
  securitySchemes:
    apiKey:
      in: header
      name: APIKEY
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http